Login+Signup+navbar+userAPI+seprateNotes

+++ backend/src/controllers/noteController.ts
@@ -3,10 +3,15 @@ import { RequestHandler } from "express";
 import NoteModel from "../models/note";
 import createHttpError from "http-errors";
 import mongoose from "mongoose";
+import { assertIsDefined } from "../util/assertIsDefined";
 
 export const getNotes: RequestHandler = async (req, res, next) => {
+
+  const authenticatedUserId = req.session.userId;
+
   try {
-    const notes = await NoteModel.find().exec();
+    assertIsDefined(authenticatedUserId);
+    const notes = await NoteModel.find({ userId: authenticatedUserId }).exec();
     res.status(200).json(notes);
   } catch (error) {
     next(error);
@@ -15,14 +20,22 @@ export const getNotes: RequestHandler = async (req, res, next) => {
 
 export const getNote: RequestHandler = async (req, res, next) => {
   const noteId = req.params.noteId;
+  const authenticatedUserId = req.session.userId;
+
 
   try {
+    assertIsDefined(authenticatedUserId);
+
     if (!mongoose.isValidObjectId(noteId)) {
       throw createHttpError(400, "Invalid Note Id");
     }
     const note = await NoteModel.findById(noteId).exec();
     if (!note) {
       throw createHttpError(404, "Note not found");
+    }
+
+    if (!note.userId.equals(authenticatedUserId)) {
+      throw createHttpError(401, "You cannot access this note.")
     }
     res.status(200).json(note);
   } catch (error) {
@@ -43,11 +56,16 @@ export const createNotes: RequestHandler<
 > = async (req, res, next) => {
   const title = req.body.title;
   const text = req.body.text;
+  const authenticatedUserId = req.session.userId;
+
   try {
+    assertIsDefined(authenticatedUserId);
+
     if (!title) {
       throw createHttpError(400, "Note must have a title");
     }
     const newNote = await NoteModel.create({
+      userId: authenticatedUserId,
       title: title,
       text: text,
     });
@@ -65,47 +83,69 @@ interface UpdateNoteBody {
   title?: string;
   text?: string;
 }
-export const updateNotes: RequestHandler<  UpdateParams,  unknown,  UpdateNoteBody,  unknown> = async (req, res, next) => {
+export const updateNotes: RequestHandler<UpdateParams, unknown, UpdateNoteBody, unknown> = async (req, res, next) => {
   const noteId = req.params.noteId;
   const newTitle = req.body.title;
   const newText = req.body.text;
+  const authenticatedUserId = req.session.userId;
+
   try {
+    assertIsDefined(authenticatedUserId);
+
     if (!mongoose.isValidObjectId(noteId)) {
-        throw createHttpError(400, "Invalid Note Id");
-      }
-      if (!newTitle) {
-        throw createHttpError(400, "Note must have a title");
-      }
+      throw createHttpError(400, "Invalid Note Id");
+    }
+    if (!newTitle) {
+      throw createHttpError(400, "Note must have a title");
+    }
 
-      const note = await NoteModel.findById(noteId).exec();
+    const note = await NoteModel.findById(noteId).exec();
 
-      if (!note) {
-        throw createHttpError(404, "Note not found");
-      }
-      note.title = newTitle;
-      note.text = newText;
+    if (!note) {
+      throw createHttpError(404, "Note not found");
+    }
 
-      const updatedNote= await note.save();
+    if (!note.userId.equals(authenticatedUserId)) {
+      throw createHttpError(401, "You cannot access this note.")
+    }
 
-      res.status(200).json(updatedNote );
+    note.title = newTitle;
+    note.text = newText;
+
+    const updatedNote = await note.save();
+
+    res.status(200).json(updatedNote);
   } catch (error) {
-    next(error); 
+    next(error);
   }
 };
 
 
 export const deleteNotes: RequestHandler = async (req, res, next) => {
-    const noteId = req.params.noteId;
-    try {
-      if (!mongoose.isValidObjectId(noteId)) {
-        throw createHttpError(400, "Invalid Note Id");
-      }
-  
-      await NoteModel.findByIdAndDelete(noteId);
-  
-      res.sendStatus(204);
-    } catch (error) {
-      next(error);
+  const noteId = req.params.noteId;
+  const authenticatedUserId = req.session.userId;
+
+  try {
+    assertIsDefined(authenticatedUserId);
+
+    if (!mongoose.isValidObjectId(noteId)) {
+      throw createHttpError(400, "Invalid Note Id");
     }
-  };
-  
+
+    const note = await NoteModel.findById(noteId).exec();
+
+    if(!note){
+      throw createHttpError(404,"Note not found")
+    }
+
+    if (!note.userId.equals(authenticatedUserId)) {
+      throw createHttpError(401, "You cannot access this note.")
+    }
+
+    await NoteModel.findByIdAndDelete(noteId);
+
+    res.sendStatus(204);
+  } catch (error) {
+    next(error);
+  }
+};
